Added text boxes for the bg colors
They are useless for now but I'll make them change the bg color when you
go back to the editor. Also while coding this I discovered that the
stupid flagging thing my ISP is doing is only doing it on the client
now. The thing that I obviously can't just hit proceed on. If you want
the full rant and more details look in the discord.