Added a variable to disable mod loading
Technically it just disables mod running. The mods still load they just
don't get ran. Also this is so that if I ever lose all my motivation for
working on the modding api I can just disable this and the code will
still be there for if I want to continue it.